[models/repo_editor] Minimising upstream file diff
Moved StartIndexing() function to repo_editor_gin.go
